+++ src/main/java/com/jeffpeng/jmod/ClientProxy.java
@@ -1,5 +1,0 @@
-package com.jeffpeng.jmod;
-
-public class ClientProxy extends Proxy {
-	public boolean preventToolBreaking = false;
-}