+++ src/main/java/com/jeffpeng/jmod/JMOD.java
@@ -7,6 +7,8 @@ import net.minecraftforge.common.MinecraftForge;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import com.jeffpeng.jmod.asm.JMODClassTransformer;
+import com.jeffpeng.jmod.asm.JMODObfuscationHelper;
 import com.jeffpeng.jmod.crafting.ToolUnbreaker;
 import com.jeffpeng.jmod.interfaces.IExecutableObject;
 import com.jeffpeng.jmod.interfaces.IStagedObject;
@@ -24,29 +26,29 @@ import com.jeffpeng.jmod.util.actions.chisel.AddCarvingVariation;
 import com.jeffpeng.jmod.util.actions.rotarycraft.AddGrinderRecipeDescriptor;
 
 import cpw.mods.fml.common.Loader;
-import cpw.mods.fml.common.Mod;
-import cpw.mods.fml.common.Mod.EventHandler;
 import cpw.mods.fml.common.ProgressManager;
 import cpw.mods.fml.common.ProgressManager.ProgressBar;
-import cpw.mods.fml.common.SidedProxy;
 import cpw.mods.fml.common.event.FMLConstructionEvent;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLLoadCompleteEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 import cpw.mods.fml.common.event.FMLServerStartedEvent;
+import cpw.mods.fml.relauncher.IFMLLoadingPlugin;
+import cpw.mods.fml.relauncher.IFMLLoadingPlugin.MCVersion;
+import cpw.mods.fml.relauncher.IFMLLoadingPlugin.TransformerExclusions;
 
 @SuppressWarnings("deprecation")
-@Mod(modid = JMOD.MODID, version = JMOD.VERSION, name = JMOD.NAME)
-public class JMOD {
-	@SidedProxy(clientSide="com.jeffpeng.jmod.ClientProxy", serverSide="com.jeffpeng.jmod.ServerProxy")
-	public static Proxy proxy;
+@MCVersion(value="1.7.10")
+@TransformerExclusions(value={"com.jeffpeng.jmod.asm"})
+public class JMOD implements IFMLLoadingPlugin {
 	public static final String MODID = "jmod";
-	public static final String VERSION = "1.0alpha1";
+	public static final String VERSION = "@VERSION@";
 	public static final String NAME = "Javascript MOD Loader";
 	private static final GlobalConfig GLOBALCONFIG = new GlobalConfig();
 	private static boolean isServer = false;
-	private static boolean devversion;
+	private static boolean devversion = false;
+	protected JMODModContainer modcontainer;
 
 	public static ForgeDeepInterface DEEPFORGE;
 	public static final JMODLoader LOADER = new JMODLoader();
@@ -55,34 +57,24 @@ public class JMOD {
 	private static JMOD instance;
 
 	public JMOD() {
-//		File nashornjar = new File(System.getProperty("java.home") + "/lib/ext/nashorn.jar");
-//		try {
-//			((ModClassLoader)Loader.instance().getModClassLoader()).addFile(nashornjar);
-//		} catch (MalformedURLException e) {
-//			// TODO Auto-generated catch block
-//			e.printStackTrace();
-//		}
-		DEEPFORGE = new ForgeDeepInterface();
-		LOADER.start();
+		instance = this;
+		JMODObfuscationHelper.init();
+		LOADER.discoverMods();
 	}
 
 	@SuppressWarnings({ "unused"})
-	@EventHandler
-	public void construct(FMLConstructionEvent event) {
-		proxy.determineIfServer();
-		
-		/*#MARK1*/if(false){
-			devversion = true;
-		} else {
-			devversion = false;
-		}
-		
-		
+	public void on(FMLConstructionEvent event) {
+		if(event.getSide().isServer()) this.isServer = true;
+		if("@devversion@".equals("true")){devversion = true;}
+		DEEPFORGE = new ForgeDeepInterface();
+		LOADER.constructMods();
+		DEEPFORGE.lockDown();
 		
 		ProgressBar bar = ProgressManager.push("Initializing JMODs", LOADER.getModList().size());
 		for(Map.Entry<String,JMODContainer> entry : LOADER.getModList().entrySet()){
 			bar.step(entry.getValue().getName());
 			entry.getValue().getMod().on(event);
+			modcontainer.meta.description += "\n    §f"+entry.getValue().getName()+"   §b"+entry.getValue().getVersion();
 		}
 		ProgressManager.pop(bar);
 		IStagedObject.sort();
@@ -91,47 +83,46 @@ public class JMOD {
 		
 	}
 
-	@EventHandler
-	public void preInit(FMLPreInitializationEvent event) {
+	
+	public void on(FMLPreInitializationEvent event) {
 		broadcast(event);
 		runningMod = null;
 	}
 
-	@EventHandler
-	public void init(FMLInitializationEvent event) {
+	
+	public void on(FMLInitializationEvent event) {
 		broadcast(event);
 		if(GLOBALCONFIG.preventToolBreaking) 	MinecraftForge.EVENT_BUS.register(new ToolUnbreaker());
-		runningMod = null;
+		if(Loader.isModLoaded("NotEnoughItem"))	new NEI_JMODConfig();
 	}
 
-	@EventHandler
-	public void postInit(FMLPostInitializationEvent event) {
+	
+	public void on(FMLPostInitializationEvent event) {
 		broadcast(event);
 
 		if(Loader.isModLoaded("RotaryCraft"))		execute(AddGrinderRecipeDescriptor.class);
 		if(Loader.isModLoaded("chisel"))			execute(AddCarvingVariation.class);
 		Lib.patchTools();
 		Lib.patchArmor();
-		runningMod = null;
+		
 	}
 	
-	@EventHandler
-	public void loadComplete(FMLLoadCompleteEvent event){
+	
+	public void on(FMLLoadCompleteEvent event){
+		
 		broadcast(event);
+		JMOD.LOG.info("##mark");
 		execute(RemoveRecipe.class);
 		execute(AddShapedRecipe.class);
 		execute(AddShapelessRecipe.class);
 		execute(AddSmeltingRecipe.class);
 		execute(RemoveChestLoot.class);
 		execute(AddChestLoot.class);
-		
-		
-		if(Loader.isModLoaded("NotEnoughItem"))	new NEI_JMODConfig();
 		if(Loader.isModLoaded("props"))			DecoCraftDyeFix.fix();
+		
 	}
 	
-	@EventHandler
-	public void serverStarted(FMLServerStartedEvent event){
+	public void on(FMLServerStartedEvent event){
 		execute(SetBlockProperties.class);
 	}
 	
@@ -152,7 +143,8 @@ public class JMOD {
 		if(event instanceof FMLPreInitializationEvent) 		IStagedObject.broadcast((FMLPreInitializationEvent)event); else
 		if(event instanceof FMLInitializationEvent) 		IStagedObject.broadcast((FMLInitializationEvent)event); else
 		if(event instanceof FMLPostInitializationEvent) 	IStagedObject.broadcast((FMLPostInitializationEvent)event); else
-		if(event instanceof FMLLoadCompleteEvent) 			IStagedObject.broadcast((FMLLoadCompleteEvent)event); 
+		if(event instanceof FMLLoadCompleteEvent) 			IStagedObject.broadcast((FMLLoadCompleteEvent)event); else
+		if(event instanceof FMLServerStartedEvent)	 		IStagedObject.broadcast((FMLServerStartedEvent)event);
 		
 	}
 	
@@ -175,4 +167,33 @@ public class JMOD {
 	public static boolean isDevVersion(){
 		return devversion;
 	}
+	
+	@Override
+	public String[] getASMTransformerClass() {
+		return new String[] { JMODClassTransformer.class.getName() };
+	}
+
+	@Override
+	public String getModContainerClass() {
+		return JMODModContainer.class.getName();
+	}
+
+	@Override
+	public String getSetupClass() {
+		return null;
+	}
+
+	@Override
+	public void injectData(Map<String, Object> data) {
+ 
+		
+	}
+
+	@Override
+	public String getAccessTransformerClass() {
+		// TODO Auto-generated method stub
+		return null;
+	}
+	
+	
 }