+++ src/main/java/com/jeffpeng/jmod/JMODLoader.java
@@ -27,25 +27,16 @@ public class JMODLoader {
 	private Map<String,JMODContainer> modList = new HashMap<String,JMODContainer>();
 	private static final String MODSDIRECTORY = "mods";
 	private static Gson gson = new GsonBuilder().setPrettyPrinting().create();
-	
-	
-	
-	public JMODLoader(){
-		
-	}
-	
-	public void start(){
-		discoverMods();
-		constructMods();
-		JMOD.DEEPFORGE.lockDown();
-		
-	}
+	private List<String> modids = new ArrayList<>();
 	
 	public Map<String,JMODContainer> getModList(){
 		return modList;
 	}
 	
-	private void discoverMods(){
+	protected void discoverMods(){
+		if(JMOD.DEEPFORGE != null && JMOD.DEEPFORGE.isLocked()){
+			throw new RuntimeException("Cannot add any more mods at this point.");
+		}
 		try {
 			Files.walk(Paths.get(MODSDIRECTORY)).forEach(filePath -> {
 			    if(isJmod(filePath) && !modQueue.contains(filePath)){
@@ -59,9 +50,15 @@ public class JMODLoader {
 		}
 	}
 	
-	private void constructMods(){
+	protected void constructMods(){
+		if(JMOD.DEEPFORGE.isLocked()){
+			throw new RuntimeException("Cannot add any more mods at this point.");
+			
+		}
 		
 		ProgressBar bar = ProgressManager.push("Constructing JMODs", modQueue.size());
+		
+		JMOD.LOG.info("Constructing " + modQueue.size() + " JMODs");
 		
 		for(Path entry : modQueue){
 			
@@ -74,7 +71,7 @@ public class JMODLoader {
 			try{
 				rawjson = Lib.readFile(entry, "mod.json");
 			} catch (IOException e){
-				JMOD.LOG.warn("Failed to load mod declaration from " + entry.toString() + ". This is an error of the mod's author. Skipping.");
+				JMOD.LOG.warn("[JMODLoader] Failed to load mod declaration from " + entry.toString() + ". This is an error of the mod's author. Skipping.");
 				continue;
 			}
 			
@@ -83,9 +80,15 @@ public class JMODLoader {
 			try {
 				 configdata = gson.fromJson(rawjson, JMODInfo.class);
 			} catch (JsonSyntaxException e){
-				JMOD.LOG.warn("Failed to parse JSON from " + entry.toString() + "   Probably the JSON is malformed. This is an error of the mod's author. Skipping.");
+				JMOD.LOG.warn("[JMODLoader] Failed to parse JSON from " + entry.toString() + "   Probably the JSON is malformed. This is an error of the mod's author. Skipping.");
 				continue;
 			}
+			
+			if(modids.contains(configdata.modid)){
+				JMOD.LOG.warn("[JMODLoader] The mod " + configdata.modid + " seems to be present more than once. Cannot load the same mod twice - fix it!");
+			}
+			
+			modids.add(configdata.modid);
 			
 			try {
 				newmod = new JMODContainer(new JMODRepresentation(configdata,!Files.isDirectory(entry)),entry.toFile());
@@ -97,11 +100,11 @@ public class JMODLoader {
 			
 			this.modList.put(newmod.getModId(),newmod);
 			JMOD.DEEPFORGE.addMod(newmod);
-			JMOD.LOG.info("successfully injected jmod \"" + newmod.getName()+"\"");
+			JMOD.LOG.info("[JMODLoader] successfully injected jmod \"" + newmod.getName()+"\"");
 		}
 		
 		ProgressManager.pop(bar);
-			
+		
 		
 	}