+++ src/main/java/com/jeffpeng/jmod/interfaces/IStagedObject.java
@@ -4,6 +4,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+import com.enderio.core.common.util.Log;
 import com.jeffpeng.jmod.JMOD;
 
 import cpw.mods.fml.common.ProgressManager;
@@ -12,6 +13,7 @@ import cpw.mods.fml.common.event.FMLInitializationEvent;
 import cpw.mods.fml.common.event.FMLLoadCompleteEvent;
 import cpw.mods.fml.common.event.FMLPostInitializationEvent;
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
+import cpw.mods.fml.common.event.FMLServerStartedEvent;
 
 @SuppressWarnings("deprecation")
 public interface IStagedObject extends Comparable<IStagedObject> {
@@ -29,7 +31,6 @@ public interface IStagedObject extends Comparable<IStagedObject> {
 		ProgressBar bar = ProgressManager.push("Staging PreInitialization", list.size());
 		for(IStagedObject o : list){
 			if(o.on(event)) c++;
-			JMOD.LOG.info("broadcast " + o.getClass().getSimpleName());
 			bar.step(o.getClass().getSimpleName());
 		}
 		ProgressManager.pop(bar);
@@ -74,6 +75,18 @@ public interface IStagedObject extends Comparable<IStagedObject> {
 		ProgressManager.pop(bar);
 		JMOD.LOG.info("[ObjectStager] Staged " + c + " of " + list.size() + " Objects");
 	}
+	
+	public static void broadcast(FMLServerStartedEvent event){
+		JMOD.LOG.info("[ObjectStager] Broadcasting FMLLoadCompleteEvent to ISIStagedObjects");
+		int c = 0;
+		ProgressBar bar = ProgressManager.push("Staging LoadComplete", list.size());
+		for(IStagedObject o : list){
+			if(o.on(event)) c++;
+			bar.step(o.getClass().getSimpleName());
+		}
+		ProgressManager.pop(bar);
+		JMOD.LOG.info("[ObjectStager] Staged " + c + " of " + list.size() + " Objects");
+	}
 
 	
 	default void registerAsStaged(){
@@ -95,6 +108,12 @@ public interface IStagedObject extends Comparable<IStagedObject> {
 	default boolean on(FMLLoadCompleteEvent event){
 		return false;
 	}
+	
+	default boolean on(FMLServerStartedEvent event){
+		return false;
+	}
+	
+	
 	
 	default int priority(){
 		return 0;