+++ src/main/java/com/jeffpeng/jmod/util/actions/AddArmorMaterial.java
@@ -59,12 +59,18 @@ public class AddArmorMaterial extends BasicAction {
 	
 	@Override
 	public boolean on(FMLPostInitializationEvent event){
-		log.info("[armor material patcher] Patching ArmorMaterial " + armormat.name());
-		new Reflector(armormat).set(5, reductionbase).set(6, factors).set(7, enchantability);
-		ArrayList<ItemStack> entryoredictstacks = OreDictionary.getOres(repairmaterial);
-		if(entryoredictstacks.size() > 0) 
-			armormat.customCraftingMaterial = entryoredictstacks.get(0).getItem();
-		else log.warn("[armor material patcher] the repairmaterial " + repairmaterial + " is unknown. " + armormat.name() + " will not be repairable.");
+		new Thread(new Runnable(){
+			@Override
+			public void run() {
+				log.info("[armor material patcher] Patching ArmorMaterial " + armormat.name());
+				new Reflector(armormat).set(5, reductionbase).set(6, factors).set(7, enchantability);
+				ArrayList<ItemStack> entryoredictstacks = OreDictionary.getOres(repairmaterial);
+				if(entryoredictstacks.size() > 0) 
+					armormat.customCraftingMaterial = entryoredictstacks.get(0).getItem();
+				else log.warn("[armor material patcher] the repairmaterial " + repairmaterial + " is unknown. " + armormat.name() + " will not be repairable.");
+			}
+		}).start();
+		
 		return true;
 	}