+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 
-v1.0 alpha 1
+alpha1.0.2
+- JMOD is now a core mod only
+- Improved loading performance
+- Added sanity check for StringListRecipes
+- Fixed crash when encountering missing metal blocks/ingots
+- Fixed creative tabs not ready on time
+- Fixed Action priority
+- Fixed OreGen not running concurrently
+
+
+alpha1.0.1
 
 - First Version, rewrite from the base of the si.core mod